move SelectedFile function to own file - RSC-1158
